// src/lib/core/type.ts
export interface CSSStyleObject {
  color?: string;
  background?: string;
  fontSize?: string;
  fontWeight?: string;
  padding?: string;
  borderRadius?: string;
  border?: string;
  textDecoration?: string;
  boxShadow?: string;
  marginTop?: string;
  marginBottom?: string;
  backgroundImage?: string;
  backgroundSize?: string;
  backgroundRepeat?: string;
  backgroundPosition?: string;
  lineHeight?: string;
}

export interface LogBadge {
  label: string;
  value: string;
  labelColor?: string;
  labelBg?: string;
  valueColor?: string;
  valueBg?: string;
}